reduce redundency

+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/loadbalance/impl/LinuxBrokerHostUsageImpl.java
@@ -123,7 +123,7 @@ public class LinuxBrokerHostUsageImpl implements BrokerHostUsage {
     }
 
     private double getTotalCpuLimit() {
-        return (double) (100 * Runtime.getRuntime().availableProcessors());
+        return 100 * Runtime.getRuntime().availableProcessors();
     }
 
     /**
@@ -189,7 +189,7 @@ public class LinuxBrokerHostUsageImpl implements BrokerHostUsage {
         if (overrideBrokerNicSpeedGbps.isPresent()) {
             // Use the override value as configured. Return the total max speed across all available NICs, converted
             // from Gbps into Kbps
-            return ((double) overrideBrokerNicSpeedGbps.get()) * nics.size() * 1024 * 1024;
+            return overrideBrokerNicSpeedGbps.get() * nics.size() * 1024 * 1024;
         }
 
         // Nic speed is in Mbits/s, return kbits/s
